aggregate zoekt.Stats (#18977)
Currently, we trace-log every event in `MeteredSearcher` which potentially
bloats the spans. With this change, we aggregate the stats for each zoekt pod.

I removed the following stats `stats.duration_ms` and  `rpc.latency_ms`
(depends on duration) because we don't set `duration` in Zoekt for streaming
search. We still keep track of "time to first event".